==> Datasets/Validation Datasets - PanEtAl/inWeb_inbiomap/1. Parsing.ipynb <==
# HQ, Confidence value = 1
# LQ, Confidence value = REST
import pandas as pd
import re

df = pd.read_csv("core.psimitab", sep = "\t", header = None)

p1 = []
p2 = []
score = []
for i in range(len(df)):
    p1.append(re.search('uniprotkb:.+_HUMAN', df.loc[i][2]).group().split(':')[1])
    p2.append(re.search('uniprotkb:.+_HUMAN', df.loc[i][3]).group().split(':')[1])
    score.append(df.loc[i][14].split("|")[0])

inWeb_inbiomap = {'Protein_1' : p1,
                  'Protein_2' : p2, 
                  'Score' : score}
inWeb_inbiomap = pd.DataFrame(inWeb_inbiomap)
inWeb_inbiomap

inWeb_inbiomap_hq = inWeb_inbiomap.loc[inWeb_inbiomap['Score'] == '1']
inWeb_inbiomap_lq = inWeb_inbiomap.loc[inWeb_inbiomap['Score'] != '1']

len(inWeb_inbiomap_hq) + len(inWeb_inbiomap_lq)

len(inWeb_inbiomap_hq)

len(inWeb_inbiomap_lq)

inWeb_inbiomap_hq.to_pickle('inWeb_inbiomap_hq.pkl')
inWeb_inbiomap_lq.to_pickle('inWeb_inbiomap_lq.pkl')
