==> Datasets/Validation Datasets - PanEtAl/inWeb_inbiomap/2. inWeb_sequence_dictionary creation.ipynb <==
from pyfaidx import Fasta
import pandas as pd
import json
import pickle
import numpy as np

def extract_id1(header):
    return header.split('|')[1]
def extract_id2(header):
    return header.split('|')[2]

#    Extract the .gz file from the following link:
#https://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/complete/uniprot_sprot.fasta.gz
#https://www.uniprot.org/downloads
#https://stackoverflow.com/questions/52569622/protein-sequence-from-uniprot-protein-id-python

sequences1 = Fasta('uniprot_sprot.fasta', key_function=extract_id1)
sequences2 = Fasta('uniprot_sprot.fasta', key_function=extract_id2)

with open('inWeb_inbiomap_hq.pkl', 'rb') as f:
    iihq_interactions = pickle.load(f)
    
with open('inWeb_inbiomap_lq.pkl', 'rb') as f:
    iilq_interactions = pickle.load(f)

iihq_interactions[['Protein_1', 'Protein_2']]

len(set(iihq_interactions.Protein_1) | set(iihq_interactions.Protein_2) | set(iilq_interactions.Protein_1) | set(iilq_interactions.Protein_2))

ii_proteins = list(set(iihq_interactions.Protein_1) | set(iihq_interactions.Protein_2) | set(iilq_interactions.Protein_1) | set(iilq_interactions.Protein_2))

str(sequences2[ii_proteins[0]])

inweb_dict = {}
not_found_seq = []
for protid in ii_proteins:
    if(protid in sequences1):
        inweb_dict[protid] = str(sequences1[protid])
    elif(protid in sequences2):
        inweb_dict[protid] = str(sequences2[protid])
    else:
        not_found_seq.append(protid)

len(not_found_seq)

len(inweb_dict)

with open("ii_sequences.json", "w") as outfile:
    json.dump(inweb_dict, outfile)

f = open('ii_sequences.json')
d = json.load(f)

len(d)
