==> Datasets/Validation Datasets - PanEtAl/inWeb_inbiomap/3. II-HQ-LQ Parsing.ipynb <==
import json
import pickle
import networkx as nx

f = open('ii_sequences.json')
d = json.load(f)

final_seq_dict_ii_le1200 = {k : d[k] for k in d.keys() if (len(d[k]) <= 1200 and len(d[k]) > 50)}

len(final_seq_dict_hippie_le1200)

f = open("inWeb_inbiomap_hq.pkl",'rb')
ii_hq = pickle.load(f)

f = open("inWeb_inbiomap_lq.pkl",'rb')
ii_lq = pickle.load(f)

iihqgr = nx.from_pandas_edgelist(ii_hq, 'Protein_1', 'Protein_2')
iilqgr = nx.from_pandas_edgelist(ii_lq, 'Protein_1', 'Protein_2')

iihqgr.remove_edges_from(nx.selfloop_edges(iihqgr))
iilqgr.remove_edges_from(nx.selfloop_edges(iilqgr))

print(iihqgr.number_of_edges())
print(iihqgr.number_of_nodes())
print(iilqgr.number_of_edges())
print(iilqgr.number_of_nodes())

not_including_list = list(set(list(iihqgr.nodes)) - set(final_seq_dict_ii_le1200.keys()))
all(proteins in list(iihqgr.nodes) for proteins in not_including_list)

for i in range(len(not_including_list)):
    iihqgr.remove_node(not_including_list[i])

print(iihqgr.number_of_edges())
print(iihqgr.number_of_nodes())

not_including_list = list(set(list(iilqgr.nodes)) - set(final_seq_dict_ii_le1200.keys()))
all(proteins in list(iilqgr.nodes) for proteins in not_including_list)

for i in range(len(not_including_list)):
    iilqgr.remove_node(not_including_list[i])

print(iilqgr.number_of_edges())
print(iilqgr.number_of_nodes())

seq_dict = {x : final_seq_dict_ii_le1200[x] for x in list(iihqgr.nodes)}
with open('hippiehqgr_int_and_dict.pkl', 'wb') as f:
    pickle.dump([nx.to_pandas_edgelist(iihqgr), seq_dict], f)
    
seq_dict = {x : final_seq_dict_ii_le1200[x] for x in list(iilqgr.nodes)}
with open('hippielqgr_int_and_dict.pkl', 'wb') as f:
    pickle.dump([nx.to_pandas_edgelist(iilqgr), seq_dict], f)
