==> Datasets/Validation Datasets - PanEtAl/inWeb_inbiomap/4. II fasta file creation .ipynb <==
import pandas as pd
import pickle

with open('iihqgr_int_and_dict.pkl', 'rb') as f:
    hippiehq_interactions, seq_dict = pickle.load(f)

df_pos = pd.read_pickle(str(Path().resolve().parent.parent) + "/Benchmark/PanEtAl/positive.pkl")
df_neg = pd.read_pickle(str(Path().resolve().parent.parent) + "/Benchmark/PanEtAl/negative.pkl")

total_proteins_bench = set(df_neg['Protein_1']) | set(df_pos['Protein_1']) | set(df_neg['Protein_2']
                                                                                ) | set(df_pos['Protein_2'])

total_proteins_bench = list(total_proteins_bench)
len(total_proteins_bench)

new_df = df_neg[['Protein_1', 'Protein_seq1']]
new_df.columns = ['ID', 'Sequence']

temp = df_neg[['Protein_2', 'Protein_seq2']]
temp.columns = ['ID', 'Sequence']
new_df = new_df.append(temp, ignore_index = True)

temp = df_pos[['Protein_1', 'Protein_seq1']]
temp.columns = ['ID', 'Sequence']
new_df = new_df.append(temp, ignore_index = True)
temp = df_pos[['Protein_2', 'Protein_seq2']]
temp.columns = ['ID', 'Sequence']
new_df = new_df.append(temp, ignore_index = True)

bench_protein_seq = new_df.set_index('ID')['Sequence'].to_dict()

res = {**bench_protein_seq, **seq_dict}

ofile = open("bench_iihq.fasta", "w")

for protein in res.keys():
    ofile.write(">" + protein + "\n" + res[protein] + "\n")

ofile.close()

with open('iilqgr_int_and_dict.pkl', 'rb') as f:
    hippielq_interactions, seq_dict = pickle.load(f)

res = {**bench_protein_seq, **seq_dict}
ofile = open("bench_iilq.fasta", "w")

for protein in res.keys():
    ofile.write(">" + protein + "\n" + res[protein] + "\n")

ofile.close()

# CD-Hit 40%

# HQ
http://weizhong-lab.ucsd.edu/cdhit_suite/cgi-bin/result.cgi?JOBID=1645601391

# LQ
http://weizhong-lab.ucsd.edu/cdhit_suite/cgi-bin/result.cgi?JOBID=1645601574
