==> Datasets/Validation Datasets - PanEtAl/inWeb_inbiomap/5. Parsing CD-Hit files - HQ.ipynb <==
import re
import pandas as pd
import pickle

# HQ
file1 = open('1645601391.result/1645601391.fas.1.clstr.sorted', 'r')
Lines = file1.readlines()

LineIndex = 0
cd_hit_dict = {}
while LineIndex < len(Lines):
    FF = re.search(r'^(>Cluster.*)', Lines[LineIndex])
    if FF != None:
        LineIndex += 1
        li = []
        dkey = ''
        while LineIndex < len(Lines):
            KK = re.search(r'^(.*\*)', Lines[LineIndex])
            if KK != None:
                dkey = KK.group()
                dkey = re.sub(r'.*>','',dkey)
                dkey = re.sub(r'\.\.\. \*','',dkey)
            else:
                KK = re.search(r'^(.*\.\.\.)', Lines[LineIndex])
                dval = KK.group()
                dval = re.sub(r'.*>','',dval)
                dval = re.sub(r'\.\.\.','',dval)
                li.append(dval)
            if(LineIndex == len(Lines)-1):
                break
            FF = re.search(r'^(>Cluster.*)', Lines[LineIndex + 1])
            if (FF == None):
                LineIndex += 1
            else:
                break
        cd_hit_dict.update({dkey: li})
    LineIndex += 1

iihq_li = []
for k in cd_hit_dict.keys():
    if(k.endswith('_HUMAN')):
        if(not(cd_hit_dict[k])):
            iihq_li.append(k)
        elif(all(protein.endswith('_HUMAN') for protein in cd_hit_dict[k])):
            iihq_li.append(k)
            iihq_li.extend(cd_hit_dict[k])
len(iihq_li)

with open('iihqgr_int_and_dict.pkl', 'rb') as f:
    ii_interactions, seq_dict = pickle.load(f)

ii_selected = {x : seq_dict[x] for x in iihq_li}

p1 = []
p2 = []
for i in range(len(ii_interactions)):
    if ((ii_interactions.loc[i][0] in ii_selected) or (ii_interactions.loc[i][1] in ii_selected)):
        p1.append(ii_interactions.loc[i][0])
        p2.append(ii_interactions.loc[i][1])

final_ii_interactions_selected = pd.DataFrame({'Protein A': p1, 'Protein B': p2})

final_ii_interactions_selected

with open('iihqlr.pkl', 'wb') as f:
    pickle.dump([final_ii_interactions_selected, seq_dict], f)

ii_interactions
